commit modified model files

+++ run.py
@@ -15,7 +15,7 @@ model = 'MILP'
 if model == 'CPSAT':
     params = src.constants.get_params(int)
 elif model == 'MILP':
-    params = src.constants.get_params(float)
+    params = src.constants.get_params(int)
 
 solver = src.run(
     params,
